added FdTracker to Events::Manager, not favourite solution

+++ Webserv/CgiModule/ImplModule/ImplModuleExternal.cpp
@@ -51,7 +51,7 @@ void	ImplModule::enqueueRequest(Cgi::Request& request, bool isCalledFromEventLoo
 
 	timeout = internal->getTimeoutMs();
 	timeout = (timeout > m_maxTimeout) ? m_maxTimeout : timeout;	
-	timeout = (timeout < ImplModule::MIN_TIMEOUT) ? ImplModule::MIN_TIMEOUT : timeout;
+	timeout = (timeout < (unsigned int)ImplModule::MIN_TIMEOUT) ? (unsigned int)ImplModule::MIN_TIMEOUT : timeout;
 	
 	// tell the requestData where its timer is, in case of premature finish/cancelation
 	internal->setMyTimer(m_timerTracker.insert(Timer::now() + timeout, internal));
